Create file for database creation
